Suppress showing item names in swap chest
This was a suggestion by ais523 from last year's tournament. The goal is
to make the biggest-donor leaderboard a little more fair by removing a
player's marginal benefit from taking an item donated by a member of
their clan.

The items in the swap chest now appear unnamed, but this is just a
special case suppression of the normal behavior of showing the name.
Once removed, the name will display as normal.