Add menu option to show achievements matching an input string
This is a QoL feature that just occurred to me. With so many
achievements, it can be hard to pick out the one you're looking for in
the list. Why not allow the player to search among them and only show
them the matches?

New option 's' in the #achievements menu prompts for a string and then
shows achievements whose names OR descriptions match that string, case
insensitive. I felt being as broad as possible would be the best way to
go about this, rather than silently hiding achievements because the case
didn't match or because only names were being scanned.