Implement 5 new achievements and document some rejects
New achievements:
  - Talk to quest leader before turn 5000
  - Swim on the Plane of Water
  - Blow up a bag of holding
  - Lose 100 HP in one turn without using lifesaving
  - Speedrun the Planes by finishing each Plane in under 100 turns

Of these, the quirkiest is the one for losing 100 HP. Because u.uhp can
be decremented in a number of places, I implemented this by checking
when the turn counter increments whether you have lost 100 HP since the
last tick - meaning that if you lose 100 HP, but still have more
actions, the achievement won't be earned until those actions are
complete, and if you happen to use them to heal up such that your net HP
loss over the turn is less than 100, you won't get it. I don't see a way
of getting around that other than tracking the HP loss in every place in
the code you can lose HP non-fatally.

Speedrunning the Plane of Water may indeed involve swimming on the
plane, but it's not strictly necessary to, so the achievement isn't
redundant.

The quest leader before turn 5000 one is an ais523 suggestion and is
implemented without changes; his intention for this achievement was to
force people to play the early game faster than is typical. I had noted
that it is probably possible to do with dig-for-victory tactics (not
necessarily all that easy, but certainly possible).