Implement 5 more achievements and document more rejects
The new achievements are:
  - Pacify a monster by untrapping it
  - Blind a monster by hitting it in the face with a cream pie
  - Enter every vault in the game (really just the Dungeons of Doom, but
    there are no other vaults outside that branch)
  - Have all the demon lords active on the same level at the same time
  - Summon 3 djinni who don't give you a wish

As noted in a comment, the demon lords achievement is implemented where
it is for performance reasons, but a player might expect to get it the
instant the last demon lord is summoned even if it doesn't get a move to
do anything. Hopefully pointing to the wording as "have them active"
will suffice if that comes up.

Probably at least one person will hit the fallback case of The Great
Heist and earn the achievement by generating the final, vaultless level
of the dungeon after digging out every other vault on the way down.
Actually, this will be common for any player who aggressively finds and
loots every vault, since levels between Medusa and the Castle are very
unlikely to have any.

The achievement does leak a tiny bit of information - namely, if you
earned it, there are no other vaults to find, and if you think you
should get it but haven't, there is another you didn't find somewhere.
However, this isn't incredibly exploitable information, and I think it's
fine.