Split "Fashionista" achievement into 4 achievements
I believe when we first implemented this achievement, it was still an
open question on how to do an identification achievement for tools which
have lots of weird subsets. Now that that was settled last year by
having multiple achievements for different subsets, it makes sense to
do the same for armor since there are really different subsets of armor
that are interesting to identify, rather than the entire object class
being one big interesting set.