Update comments about achievement free bits and total number
Noticed I had forgotten to update this for the recent commits adding
achievements. Then I also noticed the total was way less than
the #achievements command reported in the game, because I had also
forgotten to update it when I made TNNT achievements for the vanilla
achievements.

These comments might be drifting towards obsolescence; the only use I
can think of for them now is reminding the dev to add a new hex_field to
the parser when we run out of room and expand into a new tnntachieveX
xlogfile field.