Clean up TNNT section of decl.h
The inconsistency that prompted this was I noticed some constants being
used exactly once in the code (i.e. TNNT_MFORCE_GOAL is only ever used
to check if you have been mysterious forced enough to earn the
achievement). Since there are plenty of similar achievements which don't
have any such constant, I established a standard that if the constant
would be used exactly once then we shouldn't have it, but any more than
that and we should.

I did a number of other bits of cleanup:
  - Moved declarations for tnnt_globals that weren't actually in the
    tnnt_globals_t struct, into the struct at relevant positions.
  - Removed the now unused v_achieve bitfield.
  - Established another standard that defines should come before
    constants for the same achievement, and moved variables below their
    constants in accordance with that.
  - Converted tnnt_read into a function.
  - Moved defines that do NOT pertain to anything in struct
    tnnt_globals_t below and outside of it.