Add 3 "partial ironman" style achievements
The first of these, Iron Miner, was suggested by ais523 as a high-skill
achievement, in which you must proceed from the start of the game
through the end of the Mines without backtracking. I implemented it, and
expanded on the concept to add two similar achievements for Sokoban and
Gehennom.

They're not identical - Gehennom and Sokoban only start counting once
you enter through the first level of those branches. Gehennom
additionally requires you only descend by stairs and not change levels
any other way (levelporting downwards is allowed for Iron Miner, but if
we allowed it for Gehennom, it would be trivially accomplishable since
the player has much more resources to levelport down all the way).

I'm not sure whether the Sokoban achievement is too easy, since one can
simply ignore Sokoban and come back and do it later when they're strong
enough that it's easy not to backtrack. If this proves to happen a lot,
we might want to change it so that it requires no backtracking at all.