Update "Pilgrim" to record altars on conversion and require Moloch altar
A complaint from last year's tournament is that the Pilgrim achievement
was not awarded if you stood on the final altar by converting it, and it
required you to step off and back on again to record the achievement. I
fixed it by moving the code into a function to record the new altar,
which can get called from the points in the code where the hero converts
an altar.

In the process, I adjusted the description and code to require also
visiting an altar to Moloch, because if you do convert an altar, that
was 2/3 of the way there already. It's also a little more futureproof
against the eventual 3.7 merge, where the "temple of the gods" themed
room generates fairly frequently and pretty early in the game, whereas
Moloch altars are only found in special levels (but there will always
be at least 2 in any game).