Implement achievements for finding 11 different types of shops
The various types of shops are interesting enough to warrant an
achievement for each of them, just like other types of special rooms. It
helps round out the "exploration" set of achievements.

In the design discussion about this, I noted that a lot of shop types
can be obtained by repeated visits to Minetown, but not all of them.
Also, some shops that you can technically get in Minetown are rare (book
shop and wand shop come to mind) and may take a frustrating amount of
time if you're expressly going only for the achievements.

There is no achievement for entering a lighting shop, because we already
have an achievement for talking to Izchak.

Also, I replaced the name of the unrelated achievement for identifying
all rings, "Jeweler", to avoid confusion now that there is an
achievement about visiting a jewelers. The new name is superior anyway,
I think.