Implement 4 new achievements
- Reach XL 30 (this will be an achievement in 3.7, but lots of people
  wanted it to be in TNNT now)
- Pacify an angry shopkeeper (by taming, paying them off, whatever)
- Polymorph a rock into a valuable gem
- Drink a healing potion while below 10% HP... that you didn't know the
  identity of.

Regarding the fourth one, it's possible we will need to put some further
constraint on it like "you are under 10% HP and there is a monster that
just hit you" to prevent the hero from sitting in a closet and hurting
themselves down to 10% HP. But by the time 10% is a comfortable buffer,
they are likely to have identified the healing potions.

Also, the achievement name "DeathOnAStick Protege" usurps the "Death
onna Stick" achievement name for killing 25 species. I made it "Dishing
Out Death" instead. (Considered switching the kill-50-species name
"Bring Out Your Dead" with this, but I felt it was better to keep the
achievement name unchanged for consistency with previous years, since
that one didn't have to change.)