Fix: Wizard wouldn't respawn if left in deathmatch arena
The rule of "pets can't follow you through the deathmatch portal" had a
side effect where the Wizard's off-level tracking wouldn't be updated if
he appeared on the deathmatch level and the hero subsequently left,
meaning he would no longer randomly respawn as part of harassment. I
believe (though haven't tested in a long time) that he even wouldn't
respawn on the Plane of Earth.

This fixes this by changing the rule so pets can follow you out of the
arena, but still can't go in. A separate bug from the same cause is
still present due to keepdogs() not being called inbound to the
deathmatch - if the Wizard is left alive on the level outside the
portal, he will not personally reappear from off-level in the deathmatch
level as long as you remain there - but this is obviously a lot less
abusable.

While in this part of the code, I inverted the confusing if statement,
and stuck in a 3.7 todo.