Start the deathmatch when stepping off the arena door
Some feedback from ais523 last year mentioned that the deathmatch can
have weird flavor if the NPC decides to go into the antechamber or you
back out and lure them into the antechamber, and he suggested that the
door auto-lock itself. That was a fine suggestion, except the deathmatch
triggered when you stepped onto the door, not off it. So in this commit,
I adjusted the code to trigger when you step off the door, and lock it
behind you.

This precipitated some refactoring. Despite a comment to the contrary, I
realized that there is no one place in the code which moves the hero
under all circumstances. So I moved this check out of hack.c into a
tnnt.c function. While doing that, I noticed the achievement for
checking if you successfully rushed the Castle had duplicated code in
much the same places, so I also moved that into its own function.