Make deathmatch NPC creation consistent whether generic or file-read
This refactor came about because in my dev environment, I don't have any
NPC data files and hit the code path to create a generic NPC. The NPC
thus created was a poor approximation of what a real one will look like,
in particular because it didn't get any of the "default" gear the
opponent is supposed to get. This is because it simply called mk_mplayer
and exited.

Now, the code that reads the data from the file is isolated rather than
interwoven, so it can all live in the "successfully opened the file"
code path. The mk_mplayer call does not return early, but sets up
default values for the ones that would have been read out of the file
(still treating it as if the NPC had no inventory, though.) Then, the
function is able to proceed without regard to whether the variables were
read out of a data file or set as defaults for a generic opponent.

At the same time, I removed the setting of mcloned flag on the NPC -
there's a comment about that being a useful marker for the deathmatch
opponent - but we appear to have moved to using the monster id instead,
a few years ago. I added an is_deathmatch_opponent call to the one
place where mcloned still could have had an effect, preventing them from
leaving an ordinary death drop.

This shouldn't have any effect on the actual tournament since NPCs from
the previous year are pre-stocked and load in when no one has ascended
yet, but it's a marked improvement for development.