Show items in deathmatch loot pile as fully identified
The aim of this commit is to avoid giving the player the opportunity to
make a bunch of free identifications. It was possible to look up the
dumplog of the ascended character who created the deathmatch NPC, match
the items in the dumplog with any still-unidentified ones in the pile,
and make notes of what their true identities are in the current game.
This is accepted (if degenerate) behavior with bones files, but that
doesn't mean it should happen guaranteed in the deathmatch, particularly
when the NPC's inventory is likely full of very good items that are
a big benefit to have identified.

So the solution is to insert a special case, the same as how it works
for the swap chest: transient objects after the end of the deathmatch
always show as fully identified.