Add a 3.7 TODO about Potion Roulette and update achievement count
karafruit pointed out that in 3.7, it would be really easy to get
Potion Roulette due to potion of healing changes. We'll cross that
bridge when we get there, but I wanted to make sure to stick a note in
there to remind myself to revisit it.

I also had forgotten, again, to keep the achievement count updated, so
now it's up to date again.