Move the swap chests to different locations
One issue with the first swap chest, from a design standpoint, was its
predictability. A character looking for a leg up was able to rely on it
if they could make it to the bottom of the Mines. Now it lives on a
level in the Dungeons that should be roughly as difficult to reach
with usual strategies. The goal is that instead of treating it like an
expected resources, the player will get excited when they explore enough
to discover it.

The level for the chest is chosen ahead of time to avoid special levels
and the Rogue level, and it is always placed into a vault if one exists
on that level. The intention there is to make it not something you can
reliably find just by sending lots of characters diving down through the
dungeon as fast as they can go, though if there is no vault it will just
spawn in one of the level's rooms. Early midgame characters should have
enough resources (object detection, magic mapping, digging tools, etc)
to find the chest without that much trouble.

The #tnntdebug command (only available in wizard mode and beta) now
reveals the swap chest level for ease of debugging.

The second swap chest is moved from near the start of the Valley (with
even odds that you could reach it going through either 0 or 1
graveyards) to the end (favorable generation allowing you to skip the
graveyards only happens in 1/8 of games). The idea behind this is
avoiding the hero using it as an immediate post-Castle replacement for a
wish.

We discussed various other places for the second swap chest to live:
  1. Vlad's lair. The concern there was that a hero would have little
     need for a swap chest that late in the game, and it would be used
     only to dump in late-game items for others (whereas early in
     Gehennom the player likely still has a couple gaps to fill).
  2. Asmodeus's or Baalzebub's lair. The problem with these is they can
     generate quite far down into Gehennom as well.
  3. Randomly on the level below the Valley, whether that's Asmodeus's
     lair or a filler level. This was rejected for concerns of tedium
     (hunting through a 1-wide maze to find the correct square in a game
     that already forces you to do this once to win) even if most heroes
     will have a means of object detection by then.