Move TNNT_DONOR_FILE define into config.h
With it being defined only in tnnt.c unwrapped by an ifdef, it was not
possible for the person building TNNT to have the game write donors
somewhere other than $HACKDIR/donors by defining it in the compiler
cflags. This change makes it work the same as the other TNNT files and
directories.