Remove ifdefs for established TNNT features
The point of ifdef-wrapping features is to allow different people to
build the program with and without those features at the same time. This
use case doesn't pertain to TNNT, where the server build contains the
full set of features, any devs building it want to have those same
features, and the game is not generally intended for local play. It's
not like there's someone out there who wants to build TNNT except
without the swapchest, or deathmatch, etc. Plus, the ifdefs weren't
consistently used, so attempting to compile the game with no swapchest
or whatever didn't even work.

This commit removes those ifdefs, making TNNT always build them in, like
the vast majority of TNNT code.

Other related notes:
 - I left the SERVER_LOCATION ifndef alone, but put a TODO because I
   want to flip its behavior.
 - Removed a comment in end.c about "any achievements that can only be
   scored at the end of the game go here" - there are no achievements
   that are only earned in really_done(), and with good reason; we've
   made them all trigger in-game. There shouldn't be any achievements
   that get put there in the future.