Fix: robotfindskitten level was non-teleport
Noted by several players in the last tournament. ais523 best summed it
up: it was a pointless gotcha trap for unspoiled players who attempted
to teleport away from monsters on that level and found they couldn't,
since there was no indication the level was non-teleport.

Since the challenge of the level is wandering around and bumping into
non-kitten items, there isn't a compelling reason to keep this level
non-teleport. Remove that flag.