Fix: Dungeons swap chest didn't handle bones correctly
When moving the swap chest into the Dungeons of Doom, I failed to
consider bones files entirely. This had 2 possible erroneous behaviors:
1. A player might load a bones file on the level intended for the swap
   chest but which didn't have a swap chest in its original game,
   meaning the current game would not have one.
2. A player might load a bones file which did contain the swap chest in
   its original game, on a level which was not intended to have one in
   the current game, meaning the current game would have more than one
   in the Dungeons of Doom.

This fixes those problems by removing the swap chest when creating the
bones file (for non-Valley bones; the Valley swap chest is in a static
location still and uses the old code of refreshing the chest for the
next person) and by spawning a new swap chest when loading bones, if
it's the appropriate level to do so. To that end, I refactored the code
for placing the swap chest on a Dungeons level into a function, and
called that function from the makelevel code, so now there's one source
of truth for that logic.

I also added code that will prevent the swap chest landing inside a
shop. I'm not sure how that would have worked but my assumption is the
player would be charged for the item that they removed (and maybe
charged for shop-owned items placed into it). Anyway, that shouldn't be
able to happen now.