Track achievements earned in previous games
This is a big quality of life addition I wanted to get done for 2025,
removing the annoyances of going back and forth between the game and the
website to look at the achievements you still had not earned at all
during the tournament.

The first part of this was refactoring the tnnt_achievements global to
store not only the achievement strings, but a mutable status that can be
more than off/on, with the third state being "earned in a previous
game". It is now the source of truth for tnnt_is_achieved, which I also
converted into a true function rather than leaving it as a macro. The
bitmap of tnnt achievements (renamed to achievement_bitmap because it
made no sense to have separate, semantically different globals named
tnnt_achievements and tnnt_globals.tnnt_achievements) is currently
redundant with this, kept around mainly because I'm not sure I want to
loop over 300+ achievements every time one is earned to construct the
bitmap for the temp achievements file. (The other times we need it, like
writing the xlogfile, happen once or infrequently and don't really have
a performance concern.)

The second part was handling a file that contains achievements from
previous games. This file will be loaded at the start of the game, and
must be written at the end of the game. Provisionally, I've placed it in
the same temp_achievements directory with a distinct filename pattern
(but haven't checked with K2 if this will mess up any of the server-side
scripts, so they might need their own directory).

Lastly, all of this is pointless without visualization, so I heavily
altered the #achievements command:
 - Achievements earned in a previous game in the tournament but not in
   the current game use [=] as their indicator (versus [X] for
   achievements earned in the current game).
 - Added 'E' and 'U' options to show achievements earned or unearned in
   this or any previous games. I have been using the terminology "earned
   in this tournament" to encapsulate that.
 - There are separate counts at the end for the total achievements
   earned in the game and tournament. Which one gets printed depends on
   which achievements list the player wanted to see (the 'a' option to
   show all achievements prints both).