Fix: Boss Rush achievement could be worked around with doppelgangers
Furey brought up when beta testing this achievement that when you stone
to fleshed a wished-for statue of a demon lord, it becomes a
doppelganger imitating that demon lord rather than the demon lord
itself, but this still counted towards the achievement on every turn it
got to move before the doppelganger changed into something else, leading
to a weird meta where the best strategy was to gather some demon lords
and stone to flesh statues of the rest. That's not the intent of this
achievement, which is for facing all the real demon lords at once. Fix
it by eliminating shapechangers from consideration.