Optimize #achievements by only string-matching when necessary
Furey pointed out that strstri got called around 700 times when listing
achievements, because it attempted to compute search matches even when
the player is not trying to look up achievements by string. I believe
strstri with "" as its second argument will just exit, but that might be
platform-dependent behavior and it's easy enough to add a check to avoid
calling it at all.