Announce achievements at the end of the player input loop
This is another quality-of-life feature I've wanted to implement for a
year or two. Since we added the notify_achievements option, a persistent
annoyance has been the achievement notification being printed often
before the messages that invoke it, such as:

> Achievement unlocked: "One Hit KO".  You kill the newt!

This was not possible to address simply by moving the tnnt_achieve call
after other messages are printed, on account of where some of them exist
in control flow statements or helper functions. Setting up variables to
track individual achievements as earned so they could be officially
achieved later in the control flow would have meant a fair bit of
refactoring or a lot of additional tnnt_globals.

Instead, I added a flag to the tnnt_achievements global and separated
the announcement of achievements from tnnt_achieve. Now tnnt_achieve
sets that flag, and the announcement code happens in the main gameplay
loop just before the user is prompted for more input, effectively
turning it into a system where all achievements that the player earned
since their last input get printed at once and outside the flow of other
game messages.