Add 3 new conducts: containerless, zapless, and potionless
These were all proposed by ais523 as ways of incentivizing interesting
styles of play.

Containerless conduct does not mean you can't loot containers, nor from
playing as a Rogue or Archeologist with their starting sack. It means
you can't apply a container in inventory nor pick up a container not in
inventory. This prevents you from ever having more than 52 non-loadstone
items at one time, protecting your carried items from elemental or water
damage, and getting weight reduction from a bag of holding. Allowing
looting of containers means that the swap chest remains available and
players don't have to resort to bad workarounds for random chests such
as locking and bashing them until they break apart. Roles that start
with sacks still get a small benefit from them, in that they can choose
where to place a container and put items into it rather than leaving
them in piles on the ground. Currently it is allowed to obtain a bag in
inventory by wishing, or taking one out of the swap chest, because I do
not see any real way to abuse the conduct with it. All you could do with
a wished-for bag of holding is the same thing a Rogue can do with their
starting sack.

Zapless conduct bans the use of both the 'z' and 'Z' commands for
anything beyond declining to pick a wand to zap or a spell to cast. It's
worded as "attempted to zap", so a cursed wand exploding, a directional
wand selected followed by an escape out of the direction prompt ("glows
then fades"), or a forgotten spell causing confusion do break the
conduct. The intent for this is to create a game in which there isn't
magic that can be spammed repeatedly. However, other uses of wands such
as engraving and breaking are permitted under the conduct. This does
make one specific sort of magic use, burning or carving Elbereth with
wands, repeatable, but when I discussed it with ais523 he pointed out
that we already cover this with Elberethless conduct.

Potionless conduct bans quaffing, throwing, and dipping into potions. It
was proposed as a counterpart to illiterate conduct that promotes
low-resource games that yet don't change much about the overall strategy.

Of course, all three new conducts have a corresponding new achievement
added for getting assigned the Quest with the conduct unbroken.