Display backtracking status in #tnntstats
Another suggestion from Furey is to add these, to give the player a way
of confirming that they are still capable of attaining one of the
ironman-esque achievements.

To keep clutter down, I declined to add statements of the form "you have
backtracked [after entering sokoban]" since they will be true in
people's games and will not tell them anything useful far more often
than not.