Allow missiles to be added to swap chest in quantities up to 10
Several people pointed out that the change to limit the quantity of
items put into the swap chest was intended to work along the lines of
the wish parser allowing you to wish for greater quantities of
projectiles. I had forgotten, though, that the wish parser excepts not
only ammo, but also missiles (and rocks, but that will be irrelevant for
the swap chest which won't accept them).