Fix: items put into swap chest didn't respect stack splitting
This had a couple symptoms: if you explicitly tried to put in part of a
stack, you would actually put in the entire stack; and if you had more
items in the stack than acceptable and tried to specify an acceptable
quantity, it would get rejected (because it tried to put in the entire
stack).

This traces back to 3f820de, which blocked splitting of swap chest
objects because it's a huge mess if someone tried to remove a partial
stack from the swap chest. The problem is that change also blocked
putting items in, without an apparent reason to do so. I added a
condition to allow stack splitting on insertion, and it appears to be
working fine.