Add message for swap chest rejecting items because of too-large stack
One thing that wasn't sitting quite right with me after adding the stack
size limits to the swap chest was how it would treat worthy, but too
numerous, items exactly the same as unworthy items, with the possibility
of misleading the player into thinking their item was unworthy when
nothing was wrong with it other than them attempting to put in too many.

So this commit extracts the too-many logic into its own function, checks
that _after_ checking worthiness, and prints a message that more clearly
indicates that that's the problem.