Simplify logic of determining illegal level change in Gehennom
Furey was confused when reading through this logic. It was confusing
to write, so I don't blame them. They correctly pointed out that part of
the conditional was redundant and it didn't need to involve an at_stairs
check. The reason it bothered checking at_stairs was the original intent
of this block was to catch a downwards levelport in addition to other
illegal moves, and there were some convolutions to capture the idea of
"if the downward level change is happening off-stairs, it's illegal
UNLESS they are falling AND did not anticipate the fall".

As a comment notes, though, downwards levelport is already handled
elsewhere, so this block doesn't need to take care of it.