Update some comments in tnnt_achievements.h
When I added the 3 achievements for partial new conducts, I once again
forgot to update the comment. Furey was wondering why this comment still
exists since there's nothing hardcoded in the TNNT repo that needs to be
updated when a new tnntachieve* field is added to the xlogfile, so I
documented the reason: there's a hardcoded update to make on the
scoreboard side.

I also noticed the XL 30 achievement which we implemented this year was
noted in a "postpone until 3.7" comment, so I removed it. I also changed
the achievement name to the one that was suggested in the comment, since
my initial "Maxed Out" was kind of a filler name.