Fix: clarify Pilgrim and Grand Tour achievements
Pilgrim had inconsistencies in its implementation: high altars counted
towards it, but comments and strings indicated the intent behind it was
just for non-high altars. I went back and checked the commit where I
added it, thinking the original intent was for it to be an
exploration-based achievement that you can't get by stepping on the
guaranteed high altars, but this guess was actually wrong: in the
original implementation high altars count and you can always guarantee
getting it in a game. So I fixed things to account for this, including
adding "unaligned" to the list of altars visited in #tnntstats.

Meanwhile, The Grand Tour had a comment that mentioned #tnntstats should
start showing the high altar display as soon as the Invocation is done
so the player doesn't forget to step on Moloch's high altar. The problem
is the achievement didn't require Moloch's altar and the popular
conception of "The Grand Tour" as an accomplishment doesn't either (nor
did the achievement description suggest it; it explicitly says "on
Astral"). So I removed this comment and reverted it to only showing high
altar alignments after you've stepped on one on Astral.