Fix: prevent entry into the arena without starting deathmatch
This is an issue because if it's possible to get into the arena "early",
it's quite possible to just walk up to the NPC and shoot them full of
arrows or daggers without them even getting a move.

The first way to do this is just phasing through the walls as a xorn,
and was easily enough fixed by making the arena non-passwall, as it
probably should have been to start.

The second way, though, required a whole rework of how the game detects
moving into the arena. The case Furey discovered was throwing an iron
ball and being dragged by it, which does NOT use hurtling code but
instead places you on the destination spot directly. The reason this was
a problem is because the existing code relied checking you leaving the
door's space, and it was possible with high enough Strength to throw the
ball such that you never occupied the door's space at all, according to
the game.

To do this, I went back to an idea I initially had, which is to define
the arena as a region and use NetHack's existing code for checking
region entry/exit to provide a callback. This worked (though the method
for defining the region is very hacky and will need a broader revisit
when 3.7 is out and we port TNNT to it - I learned there isn't actually
a way to define it in the special level; the REGION directive does not
create this sort of region). The only problem now was, the
ball-dragging-you code does not check regions either. Turns out this is
a vanilla bug; I created a fix (adding some extra checks of
in_out_region()) and will submit it to the DevTeam later, but for now,
it's part of this commit since it's required to register the arena entry
correctly.

The tnnt_check_arena_entry function is no longer needed, since the
region method covers the stepping and hurtling cases already. I reworked
it to keep only the part that shuts the door behind the hero (in fact it
shuts and locks every door on the level, so the assumption that there
will only be 1 still stands).