Fix: achievement status wasn't part of the save file
Reported and diagnosed by Furey, who observed the behavior of after a
save and restore, achievements were not marked as earned unless earned
in a previous game, and they needed to be re-achieved them in the
current game. This is because the tnnt_achievements array, which held
static const strings up until e8ab2ac, is not saved and had no prior
need to get saved. Adding statuses to that broke that assumption.

The suggested fix was to revert the status field and have a separate
bitmap just like achievement_bitmap, except it would store the
achievements from previous games. There are two reasons that doesn't
work, though:
  1. It ignores the needs_notification field, so in practice we'd need a
     third array or bitmap to store achievements that were just earned
     and are marked for deferred notification.
  2. I set up the status enum to be extendable in the future; it's
     possible we may at some point want to implement a "achievement
     hasn't been earned by you in this tournament, but it has been
     earned by one of your clan members" status. If we put previous-game
     achievements into an array, any extension of the status would
     require yet another array for the new status and more hardcoded bit
     math.

So instead, I moved the mutable statuses into an array of similar
structure within tnnt_globals, which avoids having to save the fixed
strings (which I have renamed from tnnt_achievements to
tnnt_ach_strings) and keeps the status-using code the same.