Change shop discovery achievement criteria to "Visit a [shop]"
A few players in beta were objecting to the "Enter a [shop]" description
of the achievements, when they were invisible and the shopkeeper barred
them from actually stepping inside the shop. As far as the game is
concerned, standing on the doorway of a room counts as entering the
room, and this provides a natural trigger point for the achievements.
(It would be more complicated and expensive to try and check if the hero
is standing inside a shop each turn.)

So rather than changing the achievement logic, I changed the description
to "Visit", which hopefully carries less of a connotation of stepping
inside it (and which was already used for the achievement to find a shop
below level 20). Also considered was "Discover a [shop]", but I rejected this
because you can informally discover a shop with e.g. monster or object
detection and didn't want players to complain that doing such doesn't
award the achievement.

I also stuck a comment in the source to make it clear if the new
language still confuses people.