Fix: impossible from unpaid item being placed into swap chest
The impossible did not trigger until the game or level was saved, which
would probably have complicated finding this bug if it popped up later.
The swap chest does not spawn in a shop, but it can fall or be
teleported into one, so it should handle shops gracefully. As one would
expect, a shopkeeper who finds a swap chest in her shop should treat it
like any other container - meaning when you take an item out, you are
charged for it (this already worked without changes) and when you put an
unpaid item in, it remains on the bill (the commit fixes this part).

Also, fix the location of the achievement for putting an item into the
swap chest to be actually when the item goes into the swap chest,
because its prior location came before some checks that could stop it
from going in such as it being a welded weapon.