Expand potionless conduct to forbid breaking a potion
A couple people pointed out that it was possible to use potions to
advantage by dropping them while levitating, which only gives vapor
effects but is still tactically useful in situations like confusing
oneself on the Planes to detect portals.

I initially added an exception making it so that would break the
conduct, but then noticed it was also possible to get the same vapor
effects by putting a potion on the ground nearby and zapping it with a
force bolt. The tricky part here is that there's no way to ban that
without also banning zapping a force bolt at a monster, killing it, and
smashing a potion it was carrying, since those both use the same code
path. (And even if you could ban one without the other, that would still
provide a way to circumvent the conduct by getting a monster to pick up
the potion and then kill them).

What I ended up deciding is to ban both types of conduct, but make it
less of an unexpected gotcha by clarifying in the #conduct list that
shattering a potion is also forbidden. This means that it's not safe for
a potionless hero to attack a monster with a force bolt, at least
without first probing the monster to see if it's carrying any potions.