Expand on comment about special-cased devteam throne(s)
raisse raised a question about a second throne on the devteam level that
is obtained by transforming a sink into one, which worked the same as
Mike Stephenson's throne and had no special effects. It was ambiguous
whether this was intended behavior or not, so I extended the preexisting
comment in that code to explain that it is intended behavior.